third_party/WebKit/LayoutTests/css3/device-adapt/opera/constrain-017-expected.txt


FAIL CSS Test: @viewport constrained - width deduced from height and aspect ratio assert_equals: expected 9091 but got 320
